@@ -13,12 +13,12 @@ import java.util.Vector;
  * this template use File | Settings | File Templates.
  */
 public class MsfFile {
-    
-    private static final Logger logger = Logger.getLogger(MsfFile.class);
 
+    private static final Logger logger = Logger.getLogger(MsfFile.class);
     private File msfFile;
     private Connection iConnection;
     private Vector<AminoAcid> iAminoAcid = new Vector<AminoAcid>();
+    private MsfVersion iMsfVersion;
 
     public MsfFile(File aMsfFile) throws ClassNotFoundException, SQLException {
         this.msfFile = aMsfFile;
@@ -38,39 +38,41 @@ public class MsfFile {
         return iConnection;
     }
 
-
-    public Vector<AminoAcid> getAminoAcids(){
-        try {
-            PreparedStatement stat = iConnection.prepareStatement("select * from AminoAcids order by AminoAcidID");
-            ResultSet rs = stat.executeQuery();
-            while (rs.next()) {
-                AminoAcid aminoAcid = new AminoAcid(rs.getInt(1), rs.getString(2), rs.getString(3), rs.getString(4), rs.getDouble(5), rs.getDouble(6), rs.getString(7));
-                iAminoAcid.add(aminoAcid);
+    public Vector<AminoAcid> getAminoAcids() {
+        if (iAminoAcid.isEmpty()) {
+            try {
+                PreparedStatement stat = iConnection.prepareStatement("select * from AminoAcids order by AminoAcidID");
+                ResultSet rs = stat.executeQuery();
+                while (rs.next()) {
+                    AminoAcid aminoAcid = new AminoAcid(rs.getInt(1), rs.getString(2), rs.getString(3), rs.getString(4), rs.getDouble(5), rs.getDouble(6), rs.getString(7));
+                    iAminoAcid.add(aminoAcid);
+                }
+                rs.close();
+                stat.close();
+                return iAminoAcid;
+            } catch (SQLException e) {
+                logger.error(e);
             }
-            rs.close();
-            stat.close();
-        } catch (SQLException e) {
-            logger.error(e);
         }
         return iAminoAcid;
     }
 
-
     public MsfVersion getVersion() {
-        MsfVersion iMsfVersion = MsfVersion.VERSION1_2;
-        try {
-            Statement stat = getConnection().createStatement();
-            ResultSet rs = stat.executeQuery("select * from SchemaInfo");
-            while (rs.next()) {
-                String lVersion = rs.getString("SoftwareVersion");
-                if (lVersion.startsWith("1.2")) {
-                    iMsfVersion = MsfVersion.VERSION1_2;
-                } else if (lVersion.startsWith("1.3")) {
-                    iMsfVersion = MsfVersion.VERSION1_3;
+        if (iMsfVersion == null) {
+            try {
+                Statement stat = getConnection().createStatement();
+                ResultSet rs = stat.executeQuery("select * from SchemaInfo");
+                while (rs.next()) {
+                    String lVersion = rs.getString("SoftwareVersion");
+                    if (lVersion.startsWith("1.2")) {
+                        iMsfVersion = MsfVersion.VERSION1_2;
+                    } else if (lVersion.startsWith("1.3")) {
+                        iMsfVersion = MsfVersion.VERSION1_3;
+                    }
                 }
+            } catch (SQLException ex) {
+                logger.error(ex);
             }
-        } catch (SQLException ex) {
-            logger.error(ex);
         }
         return iMsfVersion;
     }
