@@ -8,15 +8,15 @@ import java.sql.Connection;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
-import java.util.logging.Level;
-import java.util.logging.Logger;
+import org.apache.log4j.Logger;
+
 
 /**
  *
  * @author Davy
  */
 public class WorkFlowLowMemController {
-
+private static final Logger logger = Logger.getLogger(WorkFlowLowMemController.class);
     /**
      *
      * @param aConnection a connection to the msf file
@@ -29,16 +29,20 @@ public class WorkFlowLowMemController {
             ResultSet rs = stat.executeQuery("select * from WorkflowInfo");
             rs.next();
             iWorkFlowInfo = new WorkflowInfo(rs.getString("WorkflowName"), rs.getString("WorkflowDescription"), rs.getString("User"), rs.getString("WorkflowTemplate"), rs.getString("MachineName"));
+            rs.close();
             rs = stat.executeQuery("select * from SchemaInfo");
             rs.next();
             iWorkFlowInfo.setMsfVersionInfo(new MsfVersionInfo(rs.getInt("Version"), rs.getString("SoftwareVersion")));
             //add the messages to the workflow info
+            rs.close();
             rs = stat.executeQuery("select * from WorkflowMessages");
             while (rs.next()) {
                 iWorkFlowInfo.addMessage(new WorkflowMessage(rs.getInt("MessageID"), rs.getInt("ProcessingNodeID"), rs.getInt("ProcessingNodeNumber"), rs.getLong("Time"), rs.getInt("MessageKind"), rs.getString("Message")));
             }
+            rs.close();
+            stat.close();
         } catch (SQLException ex) {
-            Logger.getLogger(WorkFlowLowMemController.class.getName()).log(Level.SEVERE, null, ex);
+            logger.error(ex);
         }
         return iWorkFlowInfo;
     }
